+++ auditor/src/client/mod.rs
@@ -13,6 +13,7 @@ use crate::{
 };
 use chrono::{DateTime, Duration, Utc};
 use reqwest;
+use serde::Serialize;
 use urlencoding::encode;
 
 static APP_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"),);
@@ -174,6 +175,139 @@ impl Default for AuditorClientBuilder {
     }
 }
 
+#[derive(serde::Deserialize, Debug, Default, Clone)]
+pub struct DateTimeUtcWrapper(pub DateTime<Utc>);
+
+impl Serialize for DateTimeUtcWrapper {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        serializer.serialize_str(&self.0.to_rfc3339())
+    }
+}
+
+/// The 'QueryParameters' is used to build query parameters which allows to query records from
+/// the database using advanced_query function.
+#[derive(serde::Deserialize, serde::Serialize, Debug, Default, Clone)]
+pub struct QueryParameters {
+    pub start_time: Option<TimeOperator>,
+    pub stop_time: Option<TimeOperator>,
+    pub runtime: Option<TimeOperator>,
+}
+
+impl Default for QueryBuilder {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+#[derive(serde::Deserialize, Debug, Clone)]
+pub enum TimeValue {
+    Datetime(DateTimeUtcWrapper),
+    Runtime(u64),
+}
+
+impl Serialize for TimeValue {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        match self {
+            TimeValue::Datetime(datetime) => datetime.serialize(serializer),
+            TimeValue::Runtime(runtime) => runtime.serialize(serializer),
+        }
+    }
+}
+
+/// The 'TimeOperator' struct is used to specify the operators on the query parameters like (gt ->
+/// greater than, lt -> lesser than, gte -> greater than equals, lte -> lesser than equals)
+#[derive(serde::Deserialize, serde::Serialize, Debug, Default, Clone)]
+pub struct TimeOperator {
+    pub gt: Option<TimeValue>,
+    pub lt: Option<TimeValue>,
+    pub gte: Option<TimeValue>,
+    pub lte: Option<TimeValue>,
+}
+
+impl TimeOperator {
+    pub fn gt(mut self, value: TimeValue) -> Self {
+        self.gt = Some(value);
+        self
+    }
+
+    pub fn lt(mut self, value: TimeValue) -> Self {
+        self.lt = Some(value);
+        self
+    }
+
+    pub fn gte(mut self, value: TimeValue) -> Self {
+        self.gte = Some(value);
+        self
+    }
+
+    pub fn lte(mut self, value: TimeValue) -> Self {
+        self.lte = Some(value);
+        self
+    }
+}
+
+impl From<chrono::DateTime<Utc>> for TimeValue {
+    fn from(item: chrono::DateTime<Utc>) -> Self {
+        TimeValue::Datetime(DateTimeUtcWrapper(item))
+    }
+}
+
+impl From<u64> for TimeValue {
+    fn from(item: u64) -> Self {
+        TimeValue::Runtime(item)
+    }
+}
+
+/// The Builder to automatically build the 'QueryParameter' struct using builder pattern
+#[derive(Debug, Clone)]
+pub struct QueryBuilder {
+    pub query_params: QueryParameters,
+}
+
+impl QueryBuilder {
+    pub fn new() -> Self {
+        QueryBuilder {
+            query_params: QueryParameters {
+                start_time: None,
+                stop_time: None,
+                runtime: None,
+            },
+        }
+    }
+
+    pub fn with_start_time(mut self, time_operator: TimeOperator) -> Self {
+        self.query_params.start_time = Some(time_operator);
+        self
+    }
+
+    pub fn with_stop_time(mut self, time_operator: TimeOperator) -> Self {
+        self.query_params.stop_time = Some(time_operator);
+        self
+    }
+
+    pub fn with_runtime(mut self, time_operator: TimeOperator) -> Self {
+        self.query_params.runtime = Some(time_operator);
+        self
+    }
+
+    pub async fn get(&self, client: AuditorClient) -> Result<Vec<Record>, ClientError> {
+        let query_string =
+            serde_qs::to_string(&self.query_params).expect("Failed to serialize query parameters");
+        println!("{}", &query_string);
+        client.advanced_query(query_string).await
+    }
+
+    pub fn build(&self) -> String {
+        serde_qs::to_string(&self.query_params).expect("Failed to serialize query parameters")
+    }
+}
+
 /// The `AuditorClient` handles the interaction with the Auditor instances and allows one to add
 /// records to the database, update records in the database and retrieve the records from the
 /// database.
@@ -285,7 +419,7 @@ impl AuditorClient {
         Ok(self
             .client
             .get(&format!(
-                "{}/record?state=started&since={}",
+                "{}/record?start_time[gte]={}",
                 &self.address, encoded_since
             ))
             .send()
@@ -314,9 +448,24 @@ impl AuditorClient {
         Ok(self
             .client
             .get(&format!(
-                "{}/record?state=stopped&since={}",
+                "{}/record?stop_time[gte]={}",
                 &self.address, encoded_since
             ))
+            .send()
+            .await?
+            .error_for_status()?
+            .json()
+            .await?)
+    }
+
+    #[tracing::instrument(
+        name = "Getting custom record queries from AUDITOR server.",
+        skip(self)
+    )]
+    pub async fn advanced_query(&self, query_string: String) -> Result<Vec<Record>, ClientError> {
+        Ok(self
+            .client
+            .get(&format!("{}/record?{}", &self.address, query_string))
             .send()
             .await?
             .error_for_status()?
@@ -428,7 +577,7 @@ impl AuditorClientBlocking {
         Ok(self
             .client
             .get(format!(
-                "{}/record?state=started&since={}",
+                "{}/record?start_time[gte]={}",
                 &self.address, encoded_since
             ))
             .send()?
@@ -452,9 +601,23 @@ impl AuditorClientBlocking {
         Ok(self
             .client
             .get(format!(
-                "{}/record?state=stopped&since={}",
+                "{}/record?stop_time[gte]={}",
                 &self.address, encoded_since
             ))
+            .send()?
+            .error_for_status()?
+            .json()?)
+    }
+
+    /// Get custom records using filters - start_time, stop_time, runtime_filters
+    ///
+    /// # Errors
+    ///
+    /// * [`ClientError::ReqwestError`] - If there was an error sending the HTTP request.
+    pub async fn advanced_query(&self, query_params: String) -> Result<Vec<Record>, ClientError> {
+        Ok(self
+            .client
+            .get(format!("{}/record?{}", &self.address, query_params))
             .send()?
             .error_for_status()?
             .json()?)
@@ -886,9 +1049,7 @@ mod tests {
 
         Mock::given(method("GET"))
             .and(path("/record"))
-            .and(query_param("state", "started"))
-            .and(query_param("since", "2022-08-03T09:47:00+00:00"))
-            // .and(query_param("since", "2022-08-03T09:47:00+00:00"))
+            .and(query_param("start_time[gte]", "2022-08-03T09:47:00+00:00"))
             .respond_with(ResponseTemplate::new(200).set_body_json(&body))
             .expect(1)
             .mount(&mock_server)
@@ -923,8 +1084,7 @@ mod tests {
 
         Mock::given(method("GET"))
             .and(path("/record"))
-            .and(query_param("state", "started"))
-            .and(query_param("since", "2022-08-03T09:47:00+00:00"))
+            .and(query_param("start_time[gte]", "2022-08-03T09:47:00+00:00"))
             .respond_with(ResponseTemplate::new(200).set_body_json(&body))
             .expect(1)
             .mount(&mock_server)
@@ -1005,8 +1165,7 @@ mod tests {
 
         Mock::given(method("GET"))
             .and(path("/record"))
-            .and(query_param("state", "stopped"))
-            .and(query_param("since", "2022-08-03T09:47:00+00:00"))
+            .and(query_param("stop_time[gte]", "2022-08-03T09:47:00+00:00"))
             .respond_with(ResponseTemplate::new(200).set_body_json(&body))
             .expect(1)
             .mount(&mock_server)
@@ -1041,8 +1200,7 @@ mod tests {
 
         Mock::given(method("GET"))
             .and(path("/record"))
-            .and(query_param("state", "stopped"))
-            .and(query_param("since", "2022-08-03T09:47:00+00:00"))
+            .and(query_param("stop_time[gte]", "2022-08-03T09:47:00+00:00"))
             .respond_with(ResponseTemplate::new(200).set_body_json(&body))
             .expect(1)
             .mount(&mock_server)
@@ -1109,5 +1267,225 @@ mod tests {
         .unwrap();
 
         assert_err!(response);
+    }
+
+    #[tokio::test]
+    async fn get_advanced_queries_succeeds() {
+        let mock_server = MockServer::start().await;
+        let client = AuditorClientBuilder::new()
+            .connection_string(&mock_server.uri())
+            .build()
+            .unwrap();
+
+        let body: Vec<Record> = vec![record()];
+
+        Mock::given(method("GET"))
+            .and(path("/record"))
+            .and(query_param("start_time[gte]", "2022-08-03T09:47:00+00:00"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(&body))
+            .expect(1)
+            .mount(&mock_server)
+            .await;
+
+        let datetime_utc = Utc.with_ymd_and_hms(2022, 8, 3, 9, 47, 0).unwrap();
+        let response = QueryBuilder::new()
+            .with_start_time(TimeOperator::default().gte(datetime_utc.into()))
+            .get(client)
+            .await
+            .unwrap();
+
+        response
+            .into_iter()
+            .zip(body)
+            .map(|(rr, br)| assert_eq!(rr, br))
+            .count();
+    }
+
+    #[tokio::test]
+    async fn get_record_query_with_start_time_and_stop_time_succeeds() {
+        let mock_server = MockServer::start().await;
+        let client = AuditorClientBuilder::new()
+            .connection_string(&mock_server.uri())
+            .build()
+            .unwrap();
+
+        let body: Vec<Record> = vec![record()];
+
+        Mock::given(method("GET"))
+            .and(path("/record"))
+            .and(query_param("start_time[gte]", "2022-08-03T09:47:00+00:00"))
+            .and(query_param("stop_time[gte]", "2022-08-03T09:47:00+00:00"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(&body))
+            .expect(1)
+            .mount(&mock_server)
+            .await;
+
+        let datetime_utc = Utc.with_ymd_and_hms(2022, 8, 3, 9, 47, 0).unwrap();
+        let response = QueryBuilder::new()
+            .with_start_time(TimeOperator::default().gte(datetime_utc.into()))
+            .with_stop_time(TimeOperator::default().gte(datetime_utc.into()))
+            .get(client)
+            .await
+            .unwrap();
+
+        response
+            .into_iter()
+            .zip(body)
+            .map(|(rr, br)| assert_eq!(rr, br))
+            .count();
+    }
+
+    #[tokio::test]
+    async fn get_record_query_with_start_time_gte_and_start_time_lte_succeeds() {
+        let mock_server = MockServer::start().await;
+        let client = AuditorClientBuilder::new()
+            .connection_string(&mock_server.uri())
+            .build()
+            .unwrap();
+
+        let body: Vec<Record> = vec![record()];
+
+        Mock::given(method("GET"))
+            .and(path("/record"))
+            .and(query_param("start_time[gte]", "2022-08-03T09:47:00+00:00"))
+            .and(query_param("start_time[lte]", "2022-08-04T09:47:00+00:00"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(&body))
+            .expect(1)
+            .mount(&mock_server)
+            .await;
+
+        let datetime_utc_gte = Utc.with_ymd_and_hms(2022, 8, 3, 9, 47, 0).unwrap();
+        let datetime_utc_lte = Utc.with_ymd_and_hms(2022, 8, 4, 9, 47, 0).unwrap();
+        let response = QueryBuilder::new()
+            .with_start_time(
+                TimeOperator::default()
+                    .gte(datetime_utc_gte.into())
+                    .lte(datetime_utc_lte.into()),
+            )
+            .get(client)
+            .await
+            .unwrap();
+
+        response
+            .into_iter()
+            .zip(body)
+            .map(|(rr, br)| assert_eq!(rr, br))
+            .count();
+    }
+
+    #[tokio::test]
+    async fn get_record_query_with_start_time_gte_and_start_time_lte_runtime_succeeds() {
+        let mock_server = MockServer::start().await;
+        let client = AuditorClientBuilder::new()
+            .connection_string(&mock_server.uri())
+            .build()
+            .unwrap();
+
+        let body: Vec<Record> = vec![record()];
+
+        Mock::given(method("GET"))
+            .and(path("/record"))
+            .and(query_param("start_time[gte]", "2022-08-03T09:47:00+00:00"))
+            .and(query_param("start_time[lte]", "2022-08-04T09:47:00+00:00"))
+            .and(query_param("runtime[gte]", "100000"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(&body))
+            .expect(1)
+            .mount(&mock_server)
+            .await;
+
+        let datetime_utc_gte = Utc.with_ymd_and_hms(2022, 8, 3, 9, 47, 0).unwrap();
+        let datetime_utc_lte = Utc.with_ymd_and_hms(2022, 8, 4, 9, 47, 0).unwrap();
+        let runtime: u64 = 100000;
+        let response = QueryBuilder::new()
+            .with_start_time(
+                TimeOperator::default()
+                    .gte(datetime_utc_gte.into())
+                    .lte(datetime_utc_lte.into()),
+            )
+            .with_runtime(TimeOperator::default().gte(runtime.into()))
+            .get(client)
+            .await
+            .unwrap();
+
+        response
+            .into_iter()
+            .zip(body)
+            .map(|(rr, br)| assert_eq!(rr, br))
+            .count();
+    }
+
+    #[tokio::test]
+    async fn get_record_query_with_start_time_stop_time_and_runtime_succeeds() {
+        let mock_server = MockServer::start().await;
+        let client = AuditorClientBuilder::new()
+            .connection_string(&mock_server.uri())
+            .build()
+            .unwrap();
+
+        let body: Vec<Record> = vec![record()];
+
+        Mock::given(method("GET"))
+            .and(path("/record"))
+            .and(query_param("start_time[gte]", "2022-08-03T09:47:00+00:00"))
+            .and(query_param("start_time[lte]", "2022-08-04T09:47:00+00:00"))
+            .and(query_param("runtime[gte]", "100000"))
+            .respond_with(ResponseTemplate::new(200).set_body_json(&body))
+            .expect(1)
+            .mount(&mock_server)
+            .await;
+
+        let datetime_utc_gte = Utc.with_ymd_and_hms(2022, 8, 3, 9, 47, 0).unwrap();
+        let datetime_utc_lte = Utc.with_ymd_and_hms(2022, 8, 4, 9, 47, 0).unwrap();
+        let runtime_gte: u64 = 100000;
+        let runtime_lte: u64 = 200000;
+        let response = QueryBuilder::new()
+            .with_start_time(
+                TimeOperator::default()
+                    .gte(datetime_utc_gte.into())
+                    .lte(datetime_utc_lte.into()),
+            )
+            .with_stop_time(
+                TimeOperator::default()
+                    .gte(datetime_utc_gte.into())
+                    .lte(datetime_utc_lte.into()),
+            )
+            .with_runtime(
+                TimeOperator::default()
+                    .gte(runtime_gte.into())
+                    .lte(runtime_lte.into()),
+            )
+            .get(client)
+            .await
+            .unwrap();
+
+        response
+            .into_iter()
+            .zip(body)
+            .map(|(rr, br)| assert_eq!(rr, br))
+            .count();
+    }
+
+    #[tokio::test]
+    async fn get_advanced_queries_fails_on_500() {
+        let mock_server = MockServer::start().await;
+        let client = AuditorClientBuilder::new()
+            .connection_string(&mock_server.uri())
+            .build()
+            .unwrap();
+
+        Mock::given(any())
+            .respond_with(ResponseTemplate::new(500))
+            .expect(1)
+            .mount(&mock_server)
+            .await;
+
+        let datetime_utc_gte = Utc.with_ymd_and_hms(2022, 8, 3, 9, 47, 0).unwrap();
+
+        assert_err!(
+            QueryBuilder::new()
+                .with_stop_time(TimeOperator::default().gte(datetime_utc_gte.into()))
+                .get(client)
+                .await
+        );
     }
 }