+++ auditor/tests/api/main.rs
@@ -1,4 +1,5 @@
 mod add;
+mod advanced_queries;
 mod client;
 mod get;
 mod get_since;